requirements.txt
ai2thor==5.0.0
ai2thor-colab==0.1.2
asttokens==2.4.1
aws-requests-auth==0.4.3
blinker==1.9.0
botocore==1.35.68
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
git+https://github.com/openai/CLIP.git
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.9
decorator==4.4.2
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
Flask==3.1.0
fonttools==4.55.0
fsspec==2024.10.0
ftfy==6.3.1
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
ipykernel==6.29.5
ipython==8.29.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
proglog==0.1.10
progressbar2==4.5.0
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-utils==3.9.0
python-xlib==0.33
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
six==1.16.0
stack-data==0.6.3
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.0
traitlets==5.14.3
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
